+++ nvi-commons/src/testFixtures/java/no/sikt/nva/nvi/common/db/DbCandidateFixtures.java
@@ -21,46 +21,53 @@ import no.sikt.nva.nvi.test.TestUtils;
 
 public class DbCandidateFixtures {
 
-    public static DbCandidate randomCandidate() {
-      return randomCandidateBuilder(true).build();
-    }
+  public static DbCandidate randomCandidate() {
+    return randomCandidateBuilder(true).build();
+  }
 
-    public static DbCandidate.Builder randomCandidateBuilder(boolean applicable) {
-      return randomCandidateBuilder(applicable, randomUri());
-    }
+  public static DbCandidate.Builder randomCandidateBuilder(boolean applicable) {
+    return randomCandidateBuilder(applicable, randomUri());
+  }
 
-    public static DbCandidate.Builder randomCandidateBuilder(boolean applicable, URI institutionId) {
-      var creatorId = randomUri();
-      var institutionPoints = TestUtils.randomBigDecimal();
-      var instanceType = randomInstanceType().getValue();
-      return DbCandidate.builder()
-          .publicationId(randomUri())
-          .publicationBucketUri(randomUri())
-          .applicable(applicable)
-          .instanceType(instanceType)
-          .points(
-              List.of(
-                  new DbInstitutionPoints(
-                      institutionId,
-                      institutionPoints,
-                      List.of(
-                          new DbCreatorAffiliationPoints(
-                              creatorId, institutionId, institutionPoints)))))
-          .level(DbLevel.LEVEL_ONE)
-          .channelType(randomElement(ChannelType.values()))
-          .channelId(randomUri())
-          .publicationDate(new DbPublicationDate(randomString(), randomString(), randomString()))
-          .internationalCollaboration(randomBoolean())
-          .creatorCount(randomInteger())
-          .createdDate(Instant.now())
-          .modifiedDate(Instant.now())
-          .totalPoints(TestUtils.randomBigDecimal())
-          .creators(
-              List.of(
-                  DbCreator
-                      .builder()
-                      .creatorId(creatorId)
-                      .affiliations(List.of(institutionId))
-                      .build()));
-    }
+  public static DbCandidate.Builder randomCandidateBuilder(boolean applicable, URI institutionId) {
+    var creatorId = randomUri();
+    var institutionPoints = TestUtils.randomBigDecimal();
+    var instanceType = randomInstanceType().getValue();
+    return DbCandidate.builder()
+        .publicationId(randomUri())
+        .publicationBucketUri(randomUri())
+        .applicable(applicable)
+        .instanceType(instanceType)
+        .points(
+            List.of(
+                new DbInstitutionPoints(
+                    institutionId,
+                    institutionPoints,
+                    List.of(
+                        new DbCreatorAffiliationPoints(
+                            creatorId, institutionId, institutionPoints)))))
+        .level(DbLevel.LEVEL_ONE)
+        .channelType(randomElement(ChannelType.values()))
+        .channelId(randomUri())
+        .publicationDate(new DbPublicationDate(randomString(), randomString(), randomString()))
+        .internationalCollaboration(randomBoolean())
+        .creatorCount(randomInteger())
+        .createdDate(Instant.now())
+        .modifiedDate(Instant.now())
+        .totalPoints(TestUtils.randomBigDecimal())
+        .creators(
+            List.of(
+                DbCreator.builder()
+                    .creatorId(creatorId)
+                    .affiliations(List.of(institutionId))
+                    .build()));
+  }
+
+  public static DbCandidate randomCandidateWithYear(String year) {
+    return randomCandidateBuilder(true).publicationDate(publicationDate(year)).build();
+  }
+
+  private static DbPublicationDate publicationDate(String year) {
+    return new DbPublicationDate(year, null, null);
+  }
 }
